helloworld/helloworld/settings/local.py
import environ
from .base import *


env = environ.Env()
environ.Env.read_env()

DEBUG = True
ALLOWED_HOSTS= ['*']

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.postgresql',
        'NAME': env('DATABASE_NAME'),
        'USER': env('DATABASE_USER'),
        'PASSWORD': env('DATABASE_PASSWORD'),
        'HOST': env('DATABASE_HOST'),
        'PORT': env('DATABASE_PORT')
    }
}