requirements.txt
django-environ==0.4.5
Django==3.2.3
djangorestframework==3.12.4
psycopg2==2.8.6
boto3==1.9.252
botocore==1.12.252
docutils==0.15.2
jmespath==0.9.4
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==5.2.0